refactor: hoc 폴더 이동
