+++ src/hoc/withLayout.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import withHead from '@/component/hoc/withHead';
+import withHead from '@/hoc/withHead';
 
 function withLayout(
   Component: (props: any) => JSX.Element,
